+++ src/plugProjectKandoU/vsCardMgr.cpp
@@ -323,7 +323,7 @@ bool VsGame::CardMgr::usePlayerCard(int user, Game::VsGame::TekiMgr* tekiMgr)
 		CI_LOOP(IPiki)
 		{
 			Piki* piki = *IPiki;
-			if (piki->getKind() == 1 - user && piki->isAlive() && (int)piki->getHappa() != Flower) {
+			if (piki->getKind() == 1 - user && piki->isAlive() && piki->getHappa() != Flower) {
 				piki->changeHappa(Flower);
 				Vector3f vec = piki->mLeafStemOffset;
 				efx::TPkGlow2 particle;
